You can now craft multi-dimensional payloads with dimensional anomalies (#89701)
## About The Pull Request
Remember the overpriced, rare bomb core from the black market, the one
you can use to convert an area to pizza or plasma?
Looking back, I think I should've made it craftable with a dimensional
anomaly rather than simply lock it to the back market, that was a real
shame to its potential. Another thing, the power of the crafted
dimensional bomb core varies with the strength of the gibtonite and if
the grenade casings used are of large or advance release variants or
not.

Using low power gibtonite (deactivated too hastily) results in a radius
of 11 (not counting the epicenter) tiles, roughly 70% of what you would
get from the blackmarket core (17 tiles).
Using a medium power gibtonite results in a radius of 13, or 82% of 17.
Using two advanced/large grenades results in a radius of 15, 88% of 17.
Using high power gibtonite results in a radius equal to the bm standard.
Using both high power gibtonite and advanced/large grenades results in a
radius of 19, for a 11.5% increase from 17.

This is done so that using the anomaly to its fullest requires some
steps beside the refined core itself, kinda in line with other anomaly
tech but without filling the techweb with more clutter.

## Why It's Good For The Game
This gives an use to the dimensional anomaly core beside the reactive
armor and bounties. If it ever becomes problematic, it can be easily
tuned later.

## Changelog

:cl:
add: You can now craft multi-dimensional payloads with dimensional
anomalies.
/:cl:

---------

Co-authored-by: SmArtKar <44720187+SmArtKar@users.noreply.github.com>