Improved Mulligan Kit (#89740)
## About The Pull Request

Instead of just being the injector, buying the Mulligan now comes as a
box containing the mutation toxin and a "Fake Identity Kit".

The Fake Identity Kit, if used inhand by someone whose name does not
exist in crew records, creates a new identity for you composed of:
- Medical and Crew Records under your new name and appearance
- An Assistant ID Card with your new name
- An additional item which creates an announcement that you have just
arrived on the station as an Assistant, as if you just did that

There's still a couple of ways you can be caught out. I didn't issue a
PDA alongside the ID (maybe I should? not sure) for one thing, and for
another those with expert identification skills (via skillchip) will be
able to tell on close inspection that the ID card was not actually
issued by Nanotrasen.
Also eagle-eyed players might get suspicious if it's announced that you
just got in on the arrival shuttle and they see you four seconds later
in the engineering lobby. Act natural, or wait for the right moment
before triggering the announcement.

## Why It's Good For The Game

The Mulligan Injector is totally useless and nobody buys it.  
After this change it's still an extremely niche item that I don't expect
will be purchased _often_ but it at least stands up to scrutiny somewhat
better, and like... gives you an Assistant ID card if that's something
that's valuable for you to have. You could actually use it to try to
become a new person without it being trivially detected by looking at
the computer.
The kit might also have niche usage for admins, who can varedit it to
have different jobs if they want someone to be added to the crew
manifest and get announced, idk.

This PR also by virtue of that I needed it adds functionality that you
can apply the appearance of any atom to a crew manifest entry, in case
some future feature needs us to have a crew manifest entry for a guy
whose photo is of a potted plant or something.

## Changelog

:cl:
add: The Mulligan Injector now comes as a kit which can create manifest
entries for you under your new identity, as well as an Assistant ID Card
with your new name.
/:cl: