Robotact is now at the top of Borg PDA screens (#90171)
## About The Pull Request
Robotact is now at the top of borg PDAs' main screen. Same as Messenger
for crew PDAs.

![image](https://github.com/user-attachments/assets/8426371f-3d16-47f2-8ec3-52d7006158da)
## Why It's Good For The Game
Robotact is just as important to borgs as Messenger is to carbons, so we
should make it easier to find at a glance.
## Changelog
:cl:
qol: Cyborg PDAs' home screens now have Robotact at the top of the
screen.
/:cl: