Fish no longer dies when flopping on safe water. (#90024)
## About The Pull Request
I've been feeling for some time that fishes dying if placed on fishable
water (or lava/plasma for a couple ones) to be kinda lakluster and
counterintuitive. YakumoChen opening an issue about it proved I wasn't
the only one thinking that, and I'm not feeling like starting with
adding fishing stuff just yet. Not without improving the existing
content first.

The source_types file was also getting big and a bit bad to navigate
through, so I've decided to split it into half a dozen distinct files:
fishing portals, rifts, turfs, structures, mining/ruins and surgery.

## Why It's Good For The Game
This fixes #89872.
The files should also be bit more organized now.

## Changelog
:cl:
fix: Fish will no longer drown if left half-submerged in a fishable turf
on which it can normally be found. Instead it will disperse after a
while or if starving. Note that this only happens if the fish is native
to such turfs (eg. lavaloop on beach water will still die).
/:cl: