Shiny cytomons! Random color mutations for vat grown creatures. (#90072)
## About The Pull Request

This PR creates a 15% chance for any vat grown creatures to be born with
a color mutation, the more dramatic the colour shift, the rarer it is!

In the future I would like to make the mutation chance dependant on the
growing conditions and cell line, but that is for another day.

The distribution between color variants is as follows:
painted: 35% 
mutant: 35%
rare: 13%
shiny: 12%
mutant king: 5%


![bild](https://github.com/user-attachments/assets/f5f110ee-d4aa-4bbe-be3d-31d0a400194a)

## Why It's Good For The Game
Cytology requires a lot of time and game knowledge and the rewards are
not that powerful considering the time investment.

So most players do not engage with it.

This is a small way to make the activity more rewarding without simply
ramping up the power level.

It also makes it easier for the cytologist to keep his sentience
potion'ed minions apart, faciltating RP with the creatures.

And we all like something rare and beautiful, don't we folks?

## Changelog

:cl:
add: cytology color mutations!
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>