Fixes the flipped table spawners (#90518)
## About The Pull Request

They weren't actually working turns out. There was a race condition with
icon smoothing, so it needs to be flipped after the table gets smoothed.
On top of that the dir wasn't being passed properly.

## Why It's Good For The Game

Stuff that actually works!

This will allow for things like this on mapload:


![image](https://github.com/user-attachments/assets/372297d7-2a30-4d85-8e89-91d14fd584d7)


## Changelog

:cl:
fix: fixes flipped table spawners not initializing correctly
/:cl: