Silicons have 66% understanding of uncommon roundstart languages (#90642)
## About The Pull Request

Silicons have 66% understanding of uncommon roundstart languages

## Why It's Good For The Game

Joke reason: 


![image](https://github.com/user-attachments/assets/12f9ede6-d808-4b61-848a-dcc8a38df6c9)

Serious reason: I understand why the original change was made, and the
points mentioned in the PR were very valid. But at the same time, I
think silicons being unable to parse most languages is a flavor fail. My
thoughts are using partial languages as a compromise would assuage the
concerns of the original PR while still giving us some of the flavor of
omniscient AI. If you disagree (Particularly you JohnFulpWillard) feel
free to close this.

(And of course we can always tweak the number. I had the thought of
adjusting the % of given languages depending on how close they are to
common - so languages like Voltaic are much lower, ~20%.)

## Changelog

:cl: Melbert
add: Cyborgs and the AI have regained understanding of various non-human
languages, but only partially (66%).
/:cl: