Supermatter anomaly fixes/explosion scaled hallucination (#90788)
## About The Pull Request

- Some fixes for the supermatter crystal anomaly generation, ensuring
they spawn outside the chamber windows and adds the outright missing
dimensional anomalies.
- Hallucination duration is calculated based on both distance from the
SM as well as the size of the delam explosion (a modifier of 0.75x for a
small explosion, up to 1.25x for the largest delam explosion)

## Why It's Good For The Game

Bugfixes, improvement on the delam hallucination.

## Changelog

:cl: LT3
fix: Fixed supermatter anomaly generation, hallucination is based on
distance and delam explosion size
/:cl: